hack/perftest/fleetbuilder/requirements.txt
azure-storage-queue==12.15.0; python_version >= '3.9'
azure-identity==1.25.3; python_version >= '3.9'
